refactor: アイデア出し画面のリファクタリング #175
