+++ front/src/app/layout.tsx
@@ -2,8 +2,8 @@ import Footer from "@/components/layouts/Footer/Footer";
 import Header from "@/components/layouts/Header/Header";
 import Stars from "@/components/layouts/Stars/Stars";
 import { Toaster } from "@/components/ui/toaster";
-import NextAuthProvider from "@/context/AuthProvider";
-import { TouchProvider } from "@/context/TouchContext";
+import NextAuthProvider from "@/context/global/AuthProvider";
+import { TouchProvider } from "@/context/global/TouchContext";
 import { openSans, zenmaruGothic } from "@/fonts/fonts";
 import { GoogleAnalytics } from "@next/third-parties/google";
 import type { Metadata } from "next";
