+++ front/src/features/generate-theme/components/AnswerSection/AnswerSection.module.scss
@@ -1,0 +1,18 @@
+/* 回答例 */
+.aiAnswerContainer {
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: flex-end;
+  gap: 8px;
+  width: 100%;
+  margin: 16px 0 30px;
+  padding: 24px;
+  border: 4px solid var(--light-gray);
+  border-radius: 1rem;
+
+  @media screen and (min-width: 1024px) {
+    margin: 24px 0 48px;
+    border-radius: 16px;
+  }
+}
