+++ front/src/features/generate-theme/components/AnswerSection/AnswerSection.tsx
@@ -1,0 +1,9 @@
+import styles from "@/features/generate-theme/components/AnswerSection/AnswerSection.module.scss";
+
+export default function AnswerSection({
+  children,
+}: {
+  children: React.ReactNode;
+}) {
+  return <div className={styles.aiAnswerContainer}>{children}</div>;
+}
