+++ front/src/features/generate-theme/components/Counter/Counter.module.scss
@@ -1,0 +1,28 @@
+.mobileCount {
+  font-size: 16px;
+  margin-top: 48px;
+
+  span {
+    font-size: 2rem;
+    font-weight: 700;
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    color: var(--light-blue);
+
+    &::before,
+    &::after {
+      content: "";
+      background-image: url("/images/bulb.svg");
+      background-size: 3rem 3rem;
+      background-repeat: no-repeat;
+      width: 3rem;
+      height: 3rem;
+      display: inline-block;
+    }
+  }
+
+  @media screen and (min-width: 768px) {
+    display: none;
+  }
+}
