+++ front/src/features/generate-theme/components/HintSection/HintSection.module.scss
@@ -1,0 +1,102 @@
+.hint {
+  width: 100%;
+  margin: 40px 0;
+
+  @media screen and (min-width: 1280px) {
+    margin: 48px 0;
+  }
+}
+
+.decoration {
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+  gap: 8px;
+  position: absolute;
+  top: -16%;
+  left: 0;
+  width: 96px;
+  height: auto;
+
+  @media screen and (min-width: 1280px) {
+    position: relative;
+    grid-column: 1 / 2;
+    grid-row: 1 / 2;
+  }
+}
+
+.fukidashi {
+  width: 100%;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  font-size: 14px;
+  color: var(--white);
+
+  @media screen and (min-width: 1280px) {
+    grid-template-columns: 120px 1fr;
+    grid-template-rows: 1fr;
+    width: 120px;
+  }
+}
+
+.astronaut {
+  width: 100%;
+  height: auto;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+}
+
+.blobContainer {
+  width: 100%;
+  height: 100%;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  position: relative;
+
+  @media screen and (min-width: 1280px) {
+    grid-column: 2 / 3;
+    grid-row: 1 / 2;
+  }
+}
+.blobCommentContainer {
+  position: relative;
+  width: 100%;
+  height: 100%;
+}
+
+.blobComment {
+  position: absolute;
+  top: 50%;
+  left: 50%;
+  transform: translate(-50%, -50%);
+  font-size: 16px;
+  color: var(--black);
+  width: 50%;
+  max-width: 50%;
+  text-align: center;
+
+  @media screen and (min-width: 1280px) {
+    font-size: 20px;
+  }
+
+  span {
+    font-size: 1.5rem;
+    font-weight: 700;
+    color: var(--pink);
+    display: block;
+
+    @media screen and (min-width: 1280px) {
+      letter-spacing: -0.1rem;
+      font-size: 2rem;
+    }
+  }
+}
+
+.blob {
+  width: 100%;
+  height: 100%;
+}
