+++ front/src/features/generate-theme/components/MyIdeaForm/MyIdeaForm.module.scss
@@ -1,0 +1,8 @@
+.form {
+  width: 100%;
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: flex-end;
+  gap: 8px;
+}
