+++ front/src/features/generate-theme/components/MyIdeaSection/MyIdeaSection.module.scss
@@ -1,0 +1,28 @@
+.answer {
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: flex-start;
+  gap: 16px;
+  width: 100%;
+  margin-bottom: 16px;
+  font-size: 16px;
+
+  @media screen and (min-width: 1024px) {
+    font-size: 20px;
+    gap: 24px;
+    margin-bottom: 24px;
+  }
+}
+
+.answerContentContainer {
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: flex-end;
+  width: 100%;
+}
+
+.answerContent {
+  width: 98%;
+}
