+++ front/src/features/generate-theme/components/NextPerspectiveButton/NextPerspectiveButton.module.scss
@@ -1,0 +1,11 @@
+.nextButton {
+  position: relative;
+}
+
+.nextArrow {
+  position: absolute;
+  top: 50%;
+  right: 0;
+  transform: translateY(-50%);
+  margin-right: 0.5rem;
+}
