+++ front/src/features/generate-theme/components/ThemeSection/ThemeSection.tsx
@@ -1,0 +1,15 @@
+import SectionTitle from "@/components/elements/SectionTitle/SectionTitle";
+import styles from "@/features/generate-theme/components/ThemeSection/ThemeSection.module.scss";
+
+export default function ThemeSection({ theme }: { theme: string | undefined }) {
+  return (
+    <div className={styles.theme}>
+      <SectionTitle>テーマ</SectionTitle>
+      <div className={styles.themeContentContainer}>
+        <div className={styles.themeContent}>
+          <div>{theme}</div>
+        </div>
+      </div>
+    </div>
+  );
+}
