+++ front/src/features/generate-theme/components/index.ts
@@ -1,0 +1,23 @@
+import AiAnswerForm from "@/features/generate-theme/components/AiAnswerForm/AiAnswerForm";
+import AiLoadingSection from "@/features/generate-theme/components/AiLoadingSection/AiLoadingSection";
+import AnswerSection from "@/features/generate-theme/components/AnswerSection/AnswerSection";
+import Counter from "@/features/generate-theme/components/Counter/Counter";
+import FixedCounter from "@/features/generate-theme/components/FixedCounter/FixedCounter";
+import HintSection from "@/features/generate-theme/components/HintSection/HintSection";
+import MyIdeaSection from "@/features/generate-theme/components/MyIdeaSection/MyIdeaSection";
+import NextOperationSection from "@/features/generate-theme/components/NextOperationSection/NextOperationSection";
+import PerspectiveSection from "@/features/generate-theme/components/PerspectiveSection/PerspectiveSection";
+import ThemeSection from "@/features/generate-theme/components/ThemeSection/ThemeSection";
+
+export {
+  AiAnswerForm,
+  AiLoadingSection,
+  AnswerSection,
+  Counter,
+  FixedCounter,
+  HintSection,
+  MyIdeaSection,
+  NextOperationSection,
+  PerspectiveSection,
+  ThemeSection,
+};
